Break Navigation Component to Smaller Components
